+++ src/Towers/CrinkleFry.cpp
@@ -26,4 +26,6 @@ void CrinkleFry::setUpUnits(){
     shared_ptr<MeleeUnit> unitToAdd = make_shared<CrinkleFryUnit>(textLoader, eventManager, textureLoader);
     currentUnits.push_back(unitToAdd);
   }
+  //call the super version to set the position of the units
+  MeleeTower::setUpUnits();
 }